Compatibility and critical design error fix
This change fixes critical part error that prevents a screw from being inserted.
It adds a simple yet complicated assembly guide for as-is top hat mod.

This also change the back skirts to make them printable in limited volume, change on side spacing covers for simplification and updated repo introduction